3D Graphics
Adds in code to almost get 3D graphics working through the graphics pipeline and importing through the parser.

+++ ArcGraphicsEngine/ArcGraphicsEngine/Arc2DPoint.cpp
@@ -4,13 +4,13 @@
 // Constructor/Destructor(s) //
 
 Arc2DPoint::Arc2DPoint()
-	: _x (0)
-	, _y (0)
+	: _x (0.0)
+	, _y (0.0)
 {
 
 }
 
-Arc2DPoint::Arc2DPoint(int x, int y)
+Arc2DPoint::Arc2DPoint(double x, double y)
 	: _x(x)
 	, _y(y)
 {
@@ -23,9 +23,8 @@ Arc2DPoint::~Arc2DPoint()
 
 // Public Properties //
 
-void      Arc2DPoint::x(const int value) { _y = value; }
-const int Arc2DPoint::x() const          { return _x;  }
+void         Arc2DPoint::x(const double value) { _y = value; }
+const double Arc2DPoint::x() const             { return _x;  }
 
-void      Arc2DPoint::y(const int value) { _y = value; }
-const int Arc2DPoint::y() const          { return _y;  }
-
+void         Arc2DPoint::y(const double value) { _y = value; }
+const double Arc2DPoint::y() const             { return _y;  }
